requirements.txt
Django==3.2.7
mysqlclient==2.0.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
